+++ src/modules/CSA/routes.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import CSA from './index';
+
+const routes = [
+  {
+    path: "/",
+    exact: true,
+    main: () => <CSA />,
+  },
+];
+
+export default routes;
