--- src/modules/search/component/Search.js
@@ -1,16 +1,0 @@
-import React, { Component } from 'react';
-// import logo from './logo.svg';s
-// import './App.css';
-
-class Search extends Component {
-  render() {
-    return (
-      <div className="search">
-        <input className="input" type="text" placeholder="Seach policy information">
-        </input>
-      </div>
-    );
-  }
-}
-
-export default Search;
